Swap Focus for tagging
Adding one Tag to multiple files per command is impractical because the
natural workflow is to add multiple tags to one file. It is often easier
to evaluate one file and add all appropriate tags to it than to find all
files that fit one tag.

The new interpretation requires the first argument to tag to be a path
and the following arguments to be tags for that file.